# src/docs/swagger.yaml
basePath: /
definitions:
  api.Entry:
    properties:
      plugin:
        type: string
      uri:
        type: string
      uuid:
        type: string
    type: object
  api.PluginEntry:
    properties:
      name:
        type: string
    type: object
host: 127.0.0.1:8085
info:
  contact: {}
  description: This is the MindfulBytes API documentation.
  title: MindfulBytes REST API
  version: "1.0"
paths:
  /v1/plugins:
    get:
      description: List all plugins.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.PluginEntry'
            type: array
      summary: List all plugins
      tags:
      - General
  /v1/plugins/{plugin}/dates:
    get:
      description: List all dates (MM-DD) for a specific plugin.
      parameters:
      - description: Plugin
        in: path
        name: plugin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: List all dates (MM-DD) for plugin
      tags:
      - General
  /v1/plugins/{plugin}/dates/{date}:
    get:
      description: List all entries for a given date (MM-DD) and plugin.
      parameters:
      - description: Plugin
        in: path
        name: plugin
        required: true
        type: string
      - description: Date
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Entry'
            type: array
      summary: List all entries for a given date (MM-DD) and plugin
      tags:
      - General
  /v1/plugins/{plugin}/fulldates:
    get:
      description: List all dates (YYYY-MM-DD) for a specific plugin.
      parameters:
      - description: Plugin
        in: path
        name: plugin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: List all dates (YYYY-MM-DD) for plugin
      tags:
      - General
  /v1/plugins/{plugin}/fulldates/{fulldate}:
    get:
      description: List all entries for a given date (YYYY-MM-DD) and plugin.
      parameters:
      - description: Plugin
        in: path
        name: plugin
        required: true
        type: string
      - description: Date (YYYY-MM-DD)
        in: path
        name: fulldate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Entry'
            type: array
      summary: List all entries for a given date (YYYY-MM-DD) and plugin
      tags:
      - General
  /v1/plugins/{plugin}/images/{imageid}:
    get:
      description: Get image with given identifier in plugin.
      parameters:
      - description: Plugin
        in: path
        name: plugin
        required: true
        type: string
      - description: Image UUID
        in: path
        name: imageid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
      summary: Get image with given identifier in plugin
      tags:
      - General
  /v1/topics:
    get:
      description: List all registered topics.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: List all topics
      tags:
      - General
  /v1/topics/{topic}/dates:
    get:
      description: List all dates for a specific topic.
      parameters:
      - description: Topic
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: List all dates for topic
      tags:
      - General
  /v1/topics/{topic}/dates/{date}:
    get:
      description: List all entries for a given date (MM-DD) and topic.
      parameters:
      - description: Topic
        in: path
        name: topic
        required: true
        type: string
      - description: Date (MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Entry'
            type: array
      summary: List all entries for a given date (MM-DD) and topic
      tags:
      - General
  /v1/topics/{topic}/fulldates:
    get:
      description: List all dates (in the form YYYY-MM-DD) for a given topic.
      parameters:
      - description: Topic
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: List all dates (YYYY-MM-DD) for topic
      tags:
      - General
  /v1/topics/{topic}/fulldates/{fulldate}:
    get:
      description: List all entries for a given date (YYYY-MM-DD) and topic.
      parameters:
      - description: Topic
        in: path
        name: topic
        required: true
        type: string
      - description: Date (YYYY-MM-DD)
        in: path
        name: fulldate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Entry'
            type: array
      summary: List all entries for a given date (YYYY-MM-DD) and topic
      tags:
      - General
  /v1/topics/{topic}/images/random:
    get:
      description: Get random image for given topic.
      parameters:
      - description: Topic
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
      summary: Get random image for given topic
      tags:
      - General
  /v1/topics/{topic}/images/today-or-random:
    get:
      description: Get Image for given topic that was created at this day x years ago. If no image is found, a random image is picked.
      parameters:
      - description: Topic
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
      summary: Get image for given topic that was created at this day x years ago or a random image.
      tags:
      - General
swagger: "2.0"
tags:
- description: List general information.
  name: General